+++ src/opus_decoder/opus_decoder.h
@@ -38,7 +38,7 @@ class OpusDecoder : public Decoder {
     uint32_t              getAudioFileDuration() override;
     const char*           getStreamTitle() override;
     const char*           whoIsIt() override;
-    int32_t               decode(uint8_t* inbuf, int32_t* bytesLeft, int16_t* outbuf) override;
+    int32_t               decode(uint8_t* inbuf, int32_t* bytesLeft, int32_t* outbuf) override;
     void                  setRawBlockParams(uint8_t channels, uint32_t sampleRate, uint8_t BPS, uint32_t tsis, uint32_t AuDaLength) override;
     std::vector<uint32_t> getMetadataBlockPicture() override;
     const char*           arg1() override;
